Added triggers gain . Requires proper testing.

+++ DrumKit/Triggers/Triggers/Trigger.h
@@ -53,17 +53,17 @@ namespace DrumKit
 
 		size_t numSamples;
 
-		short prevValue;
-		short filteredValue;
-		short prevFilteredValue;
+		double prevValue;
+		double filteredValue;
+		double prevFilteredValue;
 
 		bool trig;
 		bool out;
 
 		int64_t trigTime;
 
-		short velocity;
-		short maxVelocity;
+		double velocity;
+		double maxVelocity;
 
 		TriggerState state;
 		std::vector<curve_t<float>> curves;
